RealEstateManager

+++ app/src/main/java/com/openclassrooms/realestatemanager/model/Filter.kt
@@ -2,5 +2,5 @@ package com.openclassrooms.realestatemanager.model
 
 import java.io.Serializable
 
-data class Filter(var type: String, var priceMin: Float, var priceMax: Float, var surfaceMin: Float,
-                  var surfaceMax: Float, var interestPoint: String): Serializable
+data class Filter(var type: String, var priceMin: Int, var priceMax: Int, var surfaceMin: Int,
+                  var surfaceMax: Int, var interestPoint: String): Serializable
